+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- enable toggling top 3 priority TODOs
 
 ## [0.5.1] - 2025-10-18
 - fix issue w/ TODO reordering causing preexisting metadata to be lost